requirements.txt
pyautogui
opencv-python
Pillow
python-dotenv
mistralai